chore: updates to handle dark theme

+++ frontend/src/pages/Profile/index.js
@@ -1,13 +1,19 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState, useContext } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
+import Switch from 'react-switch';
+import { ThemeContext } from 'styled-components';
+import { shade } from 'polished';
 
+import { confirmAlert } from 'react-confirm-alert';
 import { toast } from 'react-toastify';
 import { FiPower } from 'react-icons/fi';
 
 import Button from '~/components/Button';
 import IncidentCard from '~/components/IncidentCard';
 import logo from '~/assets/logo.svg';
+import logoDark from '~/assets/logo-dark.svg';
+import Context from '~/styles/themes/context';
 
 import api from '~/services/api';
 import { signOut } from '~/store/modules/auth/actions';
@@ -15,6 +21,8 @@ import { signOut } from '~/store/modules/auth/actions';
 import { Container, Header, IncidentsList } from './styles';
 
 export default function Profile() {
+  const { colors, title } = useContext(ThemeContext);
+  const { toggleTheme } = useContext(Context);
   const history = useHistory();
   const [incidents, setIncidents] = useState([]);
   const { profile } = useSelector(state => state.ong);
@@ -23,7 +31,6 @@ export default function Profile() {
   const loadIncidents = async () => {
     try {
       const { data } = await api.get('profile');
-      console.log(data);
       setIncidents(data);
     } catch (err) {
       toast.error(
@@ -38,23 +45,49 @@ export default function Profile() {
   }, []);
 
   const handleDelete = async id => {
-    try {
-      await api.delete(`incidents/${id}`);
-      toast.success('Caso removido com sucesso');
-      setIncidents(incidents.filter(incident => incident.id !== id));
-    } catch (err) {
-      toast.error(
-        (err.response && err.response.data.error) ||
-          'Erro de comunicação com o servidor'
-      );
-    }
+    confirmAlert({
+      title: 'Confirme a exclusão',
+      message: `Deseja remover o caso ${id} ?`,
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: async () => {
+            try {
+              await api.delete(`incidents/${id}`);
+              toast.success('Caso removido com sucesso');
+              setIncidents(incidents.filter(incident => incident.id !== id));
+            } catch (err) {
+              toast.error(
+                (err.response && err.response.data.error) ||
+                  'Erro de comunicação com o servidor'
+              );
+            }
+          },
+        },
+        {
+          label: 'No',
+          onClick: () => '',
+        },
+      ],
+    });
   };
 
   return (
     <Container>
       <Header>
-        <img src={logo} alt="Heroes" />
+        <img src={title === 'light' ? logo : logoDark} alt="Heroes" />
         <span>Bem vinda, {profile.name}</span>
+        <Switch
+          onChange={toggleTheme}
+          checked={title === 'dark'}
+          height={15}
+          width={40}
+          handleDiameter={20}
+          onColor={colors.primary}
+          offColor={shade(0.3, colors.primary)}
+          onHandleColor="#fff"
+          offHandleColor="#333"
+        />
         <Button
           id="newIncident"
           type="button"
